requirements.txt
flask==2.3.2
Werkzeug==2.3.3
janome
deepl==1.18.0
python-dotenv==1.0.1